+++ src/operators/tensor/implementations/tensor_fp64x64.cairo
@@ -86,8 +86,8 @@ impl FP64x64Tensor of TensorTrait<FP64x64> {
     }
 
     fn argmax(
-        self: @Tensor<FP64x64>, axis: usize, keepdims: Option<bool>, select_last_index: Option<bool>
-    ) -> Tensor<usize> {
+        self: @Tensor<FP64x64>, axis: i32, keepdims: Option<bool>, select_last_index: Option<bool>
+    ) -> Tensor<i32> {
         math::argmax::argmax(self, axis, keepdims, select_last_index)
     }
